Finish the runnable skeleton code of leader_election client 1. launcher.sh starts the client. 2. py_http.py contains the usage.
We first use master.py to start all the servers and proxies. Then, we
kill the proxy on Bug03 to start the leader election process. After
this, we start our client to send request to Bug01. This client will
send any request to Bug01 until it's ready. It will constantly checking
the log to make sure that the leader election process is finished.